Needleman-Wunsch was broken
Michael Shapiro reported a crash in this function. While investigating
it, I figured out that some indexes were exchanged.

It is not a suprise that this function was broken for so much time. It
is not covered by regression tests. At this time, I don't add tests.
I'll do it in a future commit.