Test extension via "make installcheck"
This commit makes "make installcheck" use the actually installed
extension. In test1, we CREATE EXTENSION. In test2, we need to LOAD
'pg_similarity' because pg_regress opens a new session per test, and the
SET error is only triggered if the .so file is already loaded. test3 has
a similar problem, but here the easiest fix is to call a function first
before using SHOW.